BROKEN: Change dir structure requires two commits
Per the Flask documentation for larger applications this commit is part
one of two that moves "application.py" module into a subdirectory and
renamed __init__.py. This will allow the application to be broken in to
more modular structure and even building separate instances of the
application in the future.
Because of the way git tracks a file when it has been moved and/or
renamed this commit will break the application. Other changes to both
"application/__init__.py" and "application/config.py" are required to
bring the app back to life via the "runserver.py" module.

More specifically the UPLOAD_FOLDER variable has to be changed to
reflect the new directory structure.