feat: Add gplus oauth, not yet fully functional
The Login page now renders a template with some additional <script>s
from google as well as a bit of AJAX to execute the /gconnect view
and/or function in __init__.py. The AJAX will catch the response from
python, print it to the screen, then redirect the user to their home
screen. For the time being the login.html template is hardcoded for user
id 1, once I have the oauth issues sorted out this will be updated to
route the user dynamically based on their username, id or email.
The gconnect and gdisconnect function are not fully working. The most
recent error I've encountered is;
"TypeError: <oauth2client.client.OAuth2Credentials object at 0xb5cf296c>
is not JSON serializable"