fix: Change gconnect handling of access_token
Instead of trying to pass the whole credentials object, just use
access_token. This is a non-optimized fix however. Other ways this could
have been handled per instructor notes; to_json() method on the
credentials object OR using a very specific version of Flask inside the
vm. I'm in favor of either the fix used in this commit or the to_json()
method.