feat: Only permit the owner to edit or delete discs
The functions for adding, viewing, editing or deleting discs will now
check if the user is logged in and if they are the owner of that disc.
If they aren't logged in they will be redirected to the login page. If
they are logged in but do not own the disc they will be directed to the
public page of the disc. The new public_disc.html template is what the
public and non-owners will see of the disc.