refactor: Move show_user_home from init to views.py
This is the first of several steps to further modularize the package.
The ultimate goal is to cleanly organize, refactor and comment the code
as it leaves __init__.py and goes into views.py. To that end this
commit sees the addition of a try-except-else to safely deal with
session.querys that pull a single record that don't exist.