feat: Add Facebook login via Oauth
__init__.py now has three new functions. fbconnect(), fbdisconnect()
and just plain old disconnect(). disconnect() will do a bit of the
generic work to log out of an oauth session, making the addition of
adding new providers easier in the future. There are also new lines of
code in the login.html template to accomodate the new FB login button.